%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  zzz.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{.Last.lib}
\alias{.Last.lib}

\alias{.Last.lib}

\title{Undo changed done by this package when detached}

\usage{
.Last.lib(libpath)
}

\description{
 Undo changed done by this package when detached.
 Reverts \code{.Last()} to the function that existed before this package
 was attached.
}

\arguments{
  \item{libpath}{a character string giving the complete path to the package.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson}




\keyword{utilities}
\keyword{programming}
\keyword{internal}
