%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:09 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.binmode}
\alias{as.character.binmode}

\alias{as.character.hexmode}

\title{Converts a binary/octal/hexadecimal number into a string}

\description{
  Converts a binary/octal/hexadecimal number into a string.
}

\usage{
 \method{as.character}{binmode}(x, ...)
 \method{as.character}{hexmode}(x, ...)
}

\arguments{
 \item{x}{Object to be converted.}
  \item{...}{Not used.}
}

\value{
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{as.character.octmode()}, cf. \code{\link[base]{octmode}}.
  \code{\link{intToBin}}() (incl. \code{intToOct()} and \code{intToHex()}).
}




\keyword{manip}
\keyword{character}
\keyword{programming}
