%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiVerbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.list.MultiVerbose}
\alias{as.list.MultiVerbose}
\alias{MultiVerbose.as.list}
\alias{as.list.MultiVerbose}
\alias{as.list,MultiVerbose-method}


 \title{Gets a list of Verbose objects}

 \description{
   Gets a list of Verbose objects.
 }

 \usage{\method{as.list}{MultiVerbose}(x, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{list}} of \code{\link{Verbose}} objects.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{MultiVerbose}}.
 }
\keyword{internal}
\keyword{methods}
