%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Options.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.list.Options}
\alias{as.list.Options}
\alias{Options.as.list}
\alias{as.list.Options}
\alias{as.list,Options-method}


 \title{Gets a list representation of the options}

 \description{
  Gets a list representation of the options.
 }

 \usage{
\method{as.list}{Options}(x, ...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns a tree \code{\link[base]{list}} structure.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
