%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:31 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asByte.Java}
\alias{asByte.Java}
\alias{Java.asByte}
\alias{asByte.Java}
\alias{asByte,Java-method}


\title{Converts a numeric to a Java byte}

\description{
  Converts a numeric to a Java byte.
}

\synopsis{asByte.Java(static, x, ...)}
\usage{Java$asByte(x, ...)}

\arguments{
 \item{x}{A \code{\link[base]{numeric}} \code{\link[base]{vector}}.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Java}}.
}
\keyword{internal}
\keyword{methods}
