%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:18 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asLong.Java}
\alias{asLong.Java}
\alias{Java.asLong}
\alias{asLong.Java}
\alias{asLong,Java-method}


\title{Converts a numeric to a Java long}

\description{
  Converts a numeric to a Java long.
}

\synopsis{asLong.Java(static, x, ...)}
\usage{Java$asLong(x, ...)}

\arguments{
 \item{x}{A \code{\link[base]{numeric}} \code{\link[base]{vector}}.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Java}}.
}
\keyword{internal}
\keyword{methods}
