%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Assert.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Assert}
\docType{class}
\alias{Assert}


 \title{The Assert class}

 \description{
  Package:  R.utils \cr
\bold{Class Assert}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{Assert}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{Assert}\cr
extends \link[R.oo]{Object}\cr


 }

 \usage{
Assert(...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.utils:check.Assert]{check}} \tab Static method asserting that a generic condition is true.\cr
 \tab \code{\link[R.utils:inherits.Assert]{inherits}} \tab Static method asserting that an object inherits from of a certain class.\cr
 \tab \code{\link[R.utils:isMatrix.Assert]{isMatrix}} \tab Static method asserting thatan object is a matrix.\cr
 \tab \code{\link[R.utils:isScalar.Assert]{isScalar}} \tab Static method asserting thatan object is a single value.\cr
 \tab \code{\link[R.utils:isVector.Assert]{isVector}} \tab Static method asserting thatan object is a vector.\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save

 }

 \author{Henrik Bengtsson}
\keyword{classes}
