%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:45 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{convertComment.SmartComments}
\alias{convertComment.SmartComments}
\alias{SmartComments.convertComment}
\alias{convertComment.SmartComments}
\alias{convertComment,SmartComments-method}


 \title{Converts a single smart comment to R code}

 \description{
   Converts a single smart comment to R code.
 }

 \usage{\method{convertComment}{SmartComments}(...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Should return single \code{\link[base]{character}} of valid \R code.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{SmartComments}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
