%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  System.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{System$currentTimeMillis}
\alias{System$currentTimeMillis}
\alias{currentTimeMillis.System}
\alias{System.currentTimeMillis}
\alias{currentTimeMillis,System-method}

\title{Get the current time in milliseconds}

\usage{
## Static method (use this):
## System$currentTimeMillis(...)

## Don't use the below:
\method{currentTimeMillis}{System}(this, ...)
}

\description{
  Get the current time in milliseconds.
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[base]{Sys.time}}().
  \code{\link[base]{proc.time}}().
  For more information see \code{\link{System}}.
}
\keyword{internal}
\keyword{methods}
