%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  findSourceTraceback.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{findSourceTraceback}
\alias{findSourceTraceback.default}
\alias{findSourceTraceback}

 \title{Finds all 'srcfile' objects generated by source() in all call frames}

 \usage{\method{findSourceTraceback}{default}(...)}

 \description{
   Finds all 'srcfile' objects generated by source() in all call frames.  This makes it possible to find out which files are
   currently scripted by \code{\link[base]{source}}().
 }

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a named list of objects of class 'srcfile'.
   The names of the list entries corresponds to the 'filename'
   value of each corresponding 'srcfile' object.
   The returned list is empty if \code{\link[base]{source}}() was not called.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{IO}
\keyword{programming}
