%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  intToHex.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{format.binmode}
\alias{format.binmode}

\alias{as.character.binmode}

\title{Converts a binary/octal/hexadecimal number into a string}

\description{
  Converts a binary/octal/hexadecimal number into a string.
}

\usage{
 \method{format}{binmode}(x, ...)
}

\arguments{
 \item{x}{Object to be converted.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{format.octmode()}, cf. \code{\link[base]{octmode}}.
  \code{\link{intToBin}}() (incl. \code{intToOct()} and \code{intToHex()}).
}




\keyword{manip}
\keyword{character}
\keyword{programming}
