%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:24 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBuiltinDate.GString}
\alias{getBuiltinDate.GString}
\alias{GString.getBuiltinDate}
\alias{getBuiltinDate.GString}
\alias{getBuiltinDate,GString-method}


\title{Gets the current date}

\description{
  Gets the current date.
}

\synopsis{getBuiltinDate.GString(static, format="\%Y-\%m-\%d", ...)}
\usage{GString$getBuiltinDate(format="\%Y-\%m-\%d", ...)}

\arguments{
 \item{format}{A \code{\link[base]{character}} format string.}
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
