%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBuiltinDatetime.GString}
\alias{getBuiltinDatetime.GString}
\alias{GString.getBuiltinDatetime}
\alias{getBuiltinDatetime.GString}
\alias{getBuiltinDatetime,GString-method}


\title{Gets the current date and time}

\description{
  Gets the current date and time.
}

\usage{
## Static method (use this):
## GString$getBuiltinDatetime(format=NULL, ...)

## Don't use the below:
\method{getBuiltinDatetime}{GString}(static, format=NULL, ...)
}

\arguments{
 \item{format}{A \code{\link[base]{character}} format string.}
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
