%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TextStatusBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getLabel.TextStatusBar}
\alias{getLabel.TextStatusBar}
\alias{TextStatusBar.getLabel}
\alias{getLabel,TextStatusBar-method}

 \title{Gets the current value of a label}

 \description{
   Gets the current value of a label address either by its index or its names.
 }

 \usage{
\method{getLabel}{TextStatusBar}(this, label, ...)
}

 \arguments{
  \item{label}{The index or the name of the label.}
  \item{...}{Not used.}
 }

 \value{
   Returns the value.
 }

 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[R.utils:setLabel.TextStatusBar]{*setLabel}()} and \code{\link[R.utils:setLabels.TextStatusBar]{*setLabels}()}.
   For more information see \code{\link{TextStatusBar}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
