%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Options.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getLeaves.Options}
\alias{getLeaves.Options}
\alias{Options.getLeaves}
\alias{getLeaves,Options-method}

 \title{Gets all (non-list) options in a flat list}

 \description{
  Gets all (non-list) options in a flat list.
 }

 \usage{
\method{getLeaves}{Options}(this, ...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns a flat \code{\link[base]{list}} structure.}

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
