%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRaw.GString}
\alias{getRaw.GString}
\alias{GString.getRaw}
\alias{getRaw,GString-method}

\title{Gets the unprocessed GString}

\description{
  Gets the unprocessed GString.
}

\usage{
\method{getRaw}{GString}(object, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.utils:as.character.GString]{*as.character}()}
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
