%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:12 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRegularExpression.Arguments}
\alias{getRegularExpression.Arguments}
\alias{Arguments.getRegularExpression}
\alias{getRegularExpression.Arguments}
\alias{getRegularExpression,Arguments-method}


\title{Gets a valid regular expression pattern}

\description{
 Gets a valid regular expression pattern.
}

\synopsis{getRegularExpression.Arguments(static, pattern=NULL, ..., .name=NULL)}
\usage{Arguments$getRegularExpression(pattern=NULL, ..., .name=NULL)}

\arguments{
  \item{pattern}{A \code{\link[base]{character}} string to be validated.}
  \item{.name}{A \code{\link[base]{character}} string for name used in error messages.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[base]{grep}}().
  For more information see \code{\link{Arguments}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
