%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:27:13 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTimestampFormat.Verbose}
\alias{getTimestampFormat.Verbose}
\alias{Verbose.getTimestampFormat}
\alias{getTimestampFormat.Verbose}
\alias{getTimestampFormat,Verbose-method}


\title{Gets the default timestamp format}

\description{
  Gets the default timestamp format.
}

\usage{\method{getTimestampFormat}{Verbose}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string or a \code{\link[base]{function}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.utils:setTimestampFormat.Verbose]{*setTimestampFormat}()}.
  \code{\link[R.utils:timestampOn.Verbose]{*timestampOn}()}.
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
