%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Options.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{hasOption.Options}
\alias{hasOption.Options}
\alias{Options.hasOption}
\alias{hasOption,Options-method}

 \title{Checks if an option exists}

 \description{
  Checks if an option exists.
 }

 \usage{
\method{hasOption}{Options}(this, pathname, ...)
}

 \arguments{
   \item{pathname}{A single or a \code{\link[base]{vector}} of \code{\link[base]{character}} strings specifying
      the paths to the options to be queried.}
   \item{...}{Not used.}
 }

 \value{Returns a \code{\link[base]{logical}} (\code{\link[base]{vector}}).}

 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[R.utils:getOption.Options]{*getOption}()}.
   \code{\link[R.utils:setOption.Options]{*setOption}()}.
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
