%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  inAnyInterval.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{inAnyInterval.numeric}
\alias{inAnyInterval.numeric}
\alias{inAnyInterval.numeric}


 \title{Checks if a set of values are inside one or more intervals}

 \usage{\method{inAnyInterval}{numeric}(...)}

 \description{
  Checks if a set of values are inside one or more intervals.
 }

 \arguments{
   \item{...}{Arguments passed to \code{\link[R.utils:mapToIntervals.numeric]{*mapToIntervals}()}.}
 }

 \value{
   Returns a \code{\link[base]{logical}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
  \code{\link{mapToIntervals}}().
 }



\keyword{methods}
\keyword{utilities}
\keyword{programming}
