%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:47 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{increase.ProgressBar}
\alias{increase.ProgressBar}
\alias{ProgressBar.increase}
\alias{increase.ProgressBar}
\alias{increase,ProgressBar-method}


 \title{Increases (steps) progress bar}

 \description{
   Increases (steps) progress bar.
 }

 \usage{\method{increase}{ProgressBar}(this, stepLength=this$stepLength, visual=TRUE, ...)}

 \arguments{
  \item{stepLength}{Positive or negative step length.}
  \item{...}{Not used.}
 }

 \value{
   Returns new value.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.utils:setValue.ProgressBar]{*setValue}()}.
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
