%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:34 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isDone.ProgressBar}
\alias{isDone.ProgressBar}
\alias{ProgressBar.isDone}
\alias{isDone.ProgressBar}
\alias{isDone,ProgressBar-method}


 \title{Checks if progress bar is completed}

 \description{
   Checks if progress bar is completed.
 }

 \usage{\method{isDone}{ProgressBar}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns \code{\link[base:logical]{TRUE}} or \code{\link[base:logical]{FALSE}}.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
