%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:37 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isOn.NullVerbose}
\alias{isOn.NullVerbose}
\alias{NullVerbose.isOn}
\alias{isOn.NullVerbose}
\alias{isOn,NullVerbose-method}


\title{Checks if the output is on}

\description{
  Checks if the output is on.
}

\usage{\method{isOn}{NullVerbose}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns always \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{NullVerbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
