%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiVerbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{lapply.MultiVerbose}
\alias{lapply.MultiVerbose}
\alias{MultiVerbose.lapply}
\alias{lapply.MultiVerbose}
\alias{lapply,MultiVerbose-method}


 \title{Applies a function to each of the Verbose objects}

 \description{
   Applies a function to each of the Verbose objects.
 }

 \usage{\method{lapply}{MultiVerbose}(X, ...)}

 \arguments{
  \item{...}{Objects passed to \code{\link[base]{lapply}}().}
 }

 \value{
   Returns what \code{\link[base]{lapply}}() returns.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{MultiVerbose}}.
 }
\keyword{internal}
\keyword{methods}
