%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:19 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{LComments}
\docType{class}
\alias{LComments}


\title{The LComments class}

\description{
 Package:  R.utils \cr
\bold{Class LComments}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.utils]{SmartComments}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{\link[R.utils]{VComments}}\cr
\code{~~~~~~~~~~~~|}\cr
\code{~~~~~~~~~~~~+--}\code{LComments}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{LComments}\cr
extends \link[R.utils]{VComments}\cr



 The LComments class.

 This class, is almost identical to the super class, except that the
 constructor has different defaults.
}

\usage{LComments(letter="L", verboseName="log", ...)}

\arguments{
  \item{letter}{The smart letter.}
  \item{verboseName}{The name of the verbose object.}
  \item{...}{Not used.}
}


\author{Henrik Bengtsson \url{http://www.braju.com/R/}}



\keyword{classes}
\keyword{programming}
\keyword{IO}
