%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{less.Verbose}
\alias{less.Verbose}
\alias{Verbose.less}
\alias{less,Verbose-method}

 \title{Creates a cloned instance with a higher threshold}

 \description{
   Creates a cloned instance with a higher threshold.
 }

 \usage{
\method{less}{Verbose}(this, dThreshold=1, ...)
}

 \arguments{
  \item{dThreshold}{The amount the threshold should be raised.}
  \item{...}{Not used.}
 }

 \value{
   Returns a cloned \code{\link{Verbose}} object.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.utils:more.Verbose]{*more}()}
   For more information see \code{\link{Verbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
