%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{more.Verbose}
\alias{more.Verbose}
\alias{Verbose.more}
\alias{more.Verbose}
\alias{more,Verbose-method}


 \title{Creates a cloned instance with a lower threshold}

 \description{
   Creates a cloned instance with a lower threshold.
 }

 \usage{
\method{more}{Verbose}(this, dThreshold=1, ...)
}

 \arguments{
  \item{dThreshold}{The amount the threshold should be lowered.}
  \item{...}{Not used.}
 }

 \value{
   Returns a cloned \code{\link{Verbose}} object.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.utils:less.Verbose]{*less}()}
   For more information see \code{\link{Verbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
