%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:27 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{names.Options}
\alias{names.Options}
\alias{Options.names}
\alias{names.Options}
\alias{names,Options-method}


 \title{Gets the full pathname of all (non-list) options}

 \description{
  Gets the full pathname of all (non-list) options.
 }

 \usage{\method{names}{Options}(x, ...)}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns a \code{\link[base]{vector}} of \code{\link[base]{character}} strings.}

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
