%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:42 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfOptions.Options}
\alias{nbrOfOptions.Options}
\alias{Options.nbrOfOptions}
\alias{nbrOfOptions.Options}
\alias{nbrOfOptions,Options-method}


 \title{Gets the number of options set}

 \description{
  Gets the number of options set.
 }

 \usage{\method{nbrOfOptions}{Options}(this, ...)}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns an \code{\link[base]{integer}}.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
