%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TextStatusBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{newline.TextStatusBar}
\alias{newline.TextStatusBar}
\alias{TextStatusBar.newline}
\alias{newline,TextStatusBar-method}

 \title{Writes a newline}

 \description{
   Writes a newline.
 }

 \usage{
\method{newline}{TextStatusBar}(this, ...)
}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{TextStatusBar}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
