%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TextStatusBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{popMessage.TextStatusBar}
\alias{popMessage.TextStatusBar}
\alias{TextStatusBar.popMessage}
\alias{popMessage,TextStatusBar-method}

 \title{Adds a message above the status bar}

 \description{
   Adds a message above the status bar by scrolling up previous messages popped.
 }

 \usage{
\method{popMessage}{TextStatusBar}(this, ..., collapse="", sep="")
}

 \arguments{
  \item{...}{Arguments passed to \code{\link[base]{cat}}().}
  \item{collapse, sep}{Default values to \code{\link[base]{cat}}().}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{TextStatusBar}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
