%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{print.Verbose}
\alias{print.Verbose}
\alias{Verbose.print}
\alias{print.Verbose}
\alias{print,Verbose-method}


 \title{Prints objects if above threshold}

 \description{
   Prints objects if above threshold.
   The output is \emph{not} indented.
 }

 \usage{\method{print}{Verbose}(x, ..., level=this$defaultLevel)}

 \arguments{
  \item{...}{Objects to be passed to \code{\link[base]{print}}().}
  \item{level}{A \code{\link[base]{numeric}} value to be compared to the threshold.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
