%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:18 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{remove.FileProgressBar}
\alias{remove.FileProgressBar}
\alias{FileProgressBar.remove}
\alias{remove.FileProgressBar}
\alias{remove,FileProgressBar-method}


 \title{Removes the progress file for a file progress bar}

 \usage{\method{remove}{FileProgressBar}(object, ...)}

 \description{
   Removes the progress file for a file progress bar.
 }

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns (invisibly) \code{\link[base:logical]{TRUE}}, if there is no progress file afterwards.
   Otherwise, \code{\link[base:logical]{FALSE}} is returned.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{FileProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
