%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:36 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{reset.ProgressBar}
\alias{reset.ProgressBar}
\alias{ProgressBar.reset}
\alias{reset.ProgressBar}
\alias{reset,ProgressBar-method}


 \title{Reset progress bar}

 \description{
   Reset progress bar by setting the value to zero and updating the display.
 }

 \usage{\method{reset}{ProgressBar}(this, visual=TRUE, ...)}

 \arguments{
  \item{visual}{If \code{\link[base:logical]{TRUE}}, the progress bar is redraw, otherwise not.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   \code{\link[R.utils:setValue.ProgressBar]{*setValue}()}.
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
