%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  VComments.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{reset.VComments}
\alias{reset.VComments}
\alias{VComments.reset}
\alias{reset.VComments}
\alias{reset,VComments-method}


 \title{Resets a VComments compiler}

 \description{
   Resets a VComments compiler.
 }

 \usage{\method{reset}{VComments}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{VComments}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
