%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TextStatusBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setLabel.TextStatusBar}
\alias{setLabel.TextStatusBar}
\alias{TextStatusBar.setLabel}
\alias{setLabel,TextStatusBar-method}

 \title{Sets the value of a label}

 \description{
   Sets the value of a label address either by its index or its names.
 }

 \usage{
\method{setLabel}{TextStatusBar}(this, label, value, ...)
}

 \arguments{
  \item{label}{The index or the name of the label.}
  \item{value}{The value of the label.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[R.utils:setLabels.TextStatusBar]{*setLabels}()}
   \code{\link[R.utils:getLabel.TextStatusBar]{*getLabel}()}
   For more information see \code{\link{TextStatusBar}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
