%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TextStatusBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setLabels.TextStatusBar}
\alias{setLabels.TextStatusBar}
\alias{TextStatusBar.setLabels}
\alias{setLabels.TextStatusBar}
\alias{setLabels,TextStatusBar-method}


 \title{Sets new values of given labels}

 \description{
   Sets new values of given labels.
 }

 \usage{
\method{setLabels}{TextStatusBar}(this, ...)
}

 \arguments{
  \item{...}{A set of named arguments.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.utils:setLabel.TextStatusBar]{*setLabel}()}.
   \code{\link[R.utils:updateLabels.TextStatusBar]{*updateLabels}()}.
   For more information see \code{\link{TextStatusBar}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
