%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:34 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setMaxValue.ProgressBar}
\alias{setMaxValue.ProgressBar}
\alias{ProgressBar.setMaxValue}
\alias{setMaxValue.ProgressBar}
\alias{setMaxValue,ProgressBar-method}


 \title{Sets maximum value}

 \description{
   Sets maximum value.
 }

 \usage{\method{setMaxValue}{ProgressBar}(this, maxValue, ...)}

 \arguments{
  \item{maxValue}{New maximum value.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
