%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:47 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setTicks.ProgressBar}
\alias{setTicks.ProgressBar}
\alias{ProgressBar.setTicks}
\alias{setTicks.ProgressBar}
\alias{setTicks,ProgressBar-method}


 \title{Sets values for which ticks should be visible}

 \description{
   Sets values for which ticks should be visible.
 }

 \usage{\method{setTicks}{ProgressBar}(this, ticks=10, ...)}

 \arguments{
  \item{ticks}{Tick positions (values).}
  \item{...}{Not used.}
 }

 \value{
   Returns old tick positions.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
