%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Options.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{str.Options}
\alias{str.Options}
\alias{Options.str}
\alias{str,Options-method}

 \title{Prints the structure of the options}

 \description{
   Prints the structure of the options.
 }

 \usage{
\method{str}{Options}(object, header=paste(class(this)[1], ":\n", sep = ""), ...)
}

 \arguments{
  \item{header}{A \code{\link[base]{character}} string header to be printed at the top.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
