%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:27:12 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{str.Verbose}
\alias{str.Verbose}
\alias{Verbose.str}
\alias{str.Verbose}
\alias{str,Verbose-method}


 \title{Prints the structure of an object if above threshold}

 \description{
   Prints the structure of an object if above threshold.
   The output is \emph{not} indented.
 }

 \usage{\method{str}{Verbose}(object, ...)}

 \arguments{
  \item{...}{Objects to be passed to \code{\link[utils]{str}}.}
  \item{level}{A \code{\link[base]{numeric}} value to be compared to the threshold.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
