%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:27:13 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{timestamp.Verbose}
\alias{timestamp.Verbose}
\alias{Verbose.timestamp}
\alias{timestamp.Verbose}
\alias{timestamp,Verbose-method}


\title{Writes a timestamp}

\description{
  Writes a timestamp with default format [2005-06-23 21:20:03].
}

\usage{\method{timestamp}{Verbose}(this, format=getTimestampFormat(thos), ...)}

\arguments{
 \item{stamp}{A timestamp \code{\link[base]{character}} to be written.}
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
