%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{timestampOn.Verbose}
\alias{timestampOn.Verbose}
\alias{Verbose.timestampOn}
\alias{timestampOn,Verbose-method}
\alias{Verbose.timestampOff}
\alias{timestampOff.Verbose}
\alias{timestampOff,Verbose-method}


\title{Turns automatic timestamping on and off}

\description{
  Turns automatic timestamping on and off.
}

\usage{
\method{timestampOn}{Verbose}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns (invisibly) the old timestamp status.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.utils:setTimestampFormat.Verbose]{*setTimestampFormat}()}.
  \code{\link[R.utils:timestampOn.Verbose]{*timestampOn}()}.
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
