%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:36 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{update.ProgressBar}
\alias{update.ProgressBar}
\alias{ProgressBar.update}
\alias{update.ProgressBar}
\alias{update,ProgressBar-method}


 \title{Updates progress bar}

 \description{
   Updates progress bar.
 }

 \usage{\method{update}{ProgressBar}(object, visual=TRUE, ...)}

 \arguments{
  \item{visual}{If \code{\link[base:logical]{TRUE}}, the progress bar is redrawn, otherwise not.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
