%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TextStatusBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{updateLabels.TextStatusBar}
\alias{updateLabels.TextStatusBar}
\alias{TextStatusBar.updateLabels}
\alias{updateLabels.TextStatusBar}
\alias{updateLabels,TextStatusBar-method}


 \title{Sets the new values of given labels and updates the status bar}

 \description{
   Sets the new values of given labels and updates the status bar.
 }

 \usage{\method{updateLabels}{TextStatusBar}(this, ...)}

 \arguments{
  \item{...}{A set of named arguments.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.utils:setLabels.TextStatusBar]{*setLabels}()}.
   For more information see \code{\link{TextStatusBar}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
