%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:46 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{validate.SmartComments}
\alias{validate.SmartComments}
\alias{SmartComments.validate}
\alias{validate.SmartComments}
\alias{validate,SmartComments-method}


 \title{Validates the compiled lines}

 \description{
   Validates the compiled lines
 }

 \usage{\method{validate}{SmartComments}(this, lines, ...)}

 \arguments{
  \item{lines}{A \code{\link[base]{character}} \code{\link[base]{vector}} of lines of code to validated.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{SmartComments}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
