%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  VComments.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{validate.VComments}
\alias{validate.VComments}
\alias{VComments.validate}
\alias{validate.VComments}
\alias{validate,VComments-method}


 \title{Validates the compiled lines}

 \description{
   Validates the compiled lines
 }

 \usage{
\method{validate}{VComments}(this, lines, ...)
}

 \arguments{
  \item{lines}{A \code{\link[base]{character}} \code{\link[base]{vector}} of lines of code to validated.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{VComments}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
