%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiVerbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{writeRaw.MultiVerbose}
\alias{writeRaw.MultiVerbose}
\alias{MultiVerbose.writeRaw}
\alias{writeRaw.MultiVerbose}
\alias{writeRaw,MultiVerbose-method}


 \title{Writes to each of the Verbose objects}

 \description{
   Writes to each of the Verbose objects.
 }

 \usage{\method{writeRaw}{MultiVerbose}(this, ...)}

 \arguments{
  \item{...}{Additiona objects to be passed to \code{writeRaw()} for
     each \code{\link{Verbose}} object.}
 }

 \value{
   Returns (invisibly) \code{\link[base:logical]{TRUE}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{MultiVerbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
