%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  NullVerbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{writeRaw.NullVerbose}
\alias{writeRaw.NullVerbose}
\alias{NullVerbose.writeRaw}
\alias{writeRaw.NullVerbose}
\alias{writeRaw,NullVerbose-method}

 \alias{NullVerbose.cat}
\alias{cat.NullVerbose}
\alias{cat,NullVerbose-method}

 \alias{NullVerbose.printf}
\alias{printf.NullVerbose}
\alias{printf,NullVerbose-method}

 \alias{NullVerbose.enter}
\alias{enter.NullVerbose}
\alias{enter,NullVerbose-method}

 \alias{NullVerbose.exit}
\alias{exit.NullVerbose}
\alias{exit,NullVerbose-method}

 \alias{NullVerbose.print}
\alias{print.NullVerbose}
\alias{print,NullVerbose-method}

 \alias{NullVerbose.str}
\alias{str.NullVerbose}
\alias{str,NullVerbose-method}

 \alias{NullVerbose.summary}
\alias{summary.NullVerbose}
\alias{summary,NullVerbose-method}

 \alias{NullVerbose.evaluate}
\alias{evaluate.NullVerbose}
\alias{evaluate,NullVerbose-method}

 \alias{NullVerbose.newline}
\alias{newline.NullVerbose}
\alias{newline,NullVerbose-method}

 \alias{NullVerbose.ruler}
\alias{ruler.NullVerbose}
\alias{ruler,NullVerbose-method}

 \alias{NullVerbose.header}
\alias{header.NullVerbose}
\alias{header,NullVerbose-method}


 \title{All output methods}

 \description{
   All output methods of this class ignores their input arguments and outputs
   nothing.
 }

 \usage{
\method{writeRaw}{NullVerbose}(...)
}

 \arguments{
  \item{...}{Ignored.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{NullVerbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
