%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Arguments.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Arguments}
\docType{class}
\alias{Arguments}


\title{Static class to validate and process arguments}

\description{
 Package:  R.utils \cr
\bold{Class Arguments}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{Arguments}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{Arguments}\cr
extends \link[R.oo]{Object}\cr


}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{getCharacter} \tab  -\cr
 \tab \code{getCharacters} \tab  -\cr
 \tab \code{getDirectory} \tab  -\cr
 \tab \code{getDouble} \tab  -\cr
 \tab \code{getDoubles} \tab  -\cr
 \tab \code{getEnvironment} \tab  -\cr
 \tab \code{getIndex} \tab  -\cr
 \tab \code{getIndices} \tab  -\cr
 \tab \code{getInteger} \tab  -\cr
 \tab \code{getIntegers} \tab  -\cr
 \tab \code{getLogical} \tab  -\cr
 \tab \code{getLogicals} \tab  -\cr
 \tab \code{getNumeric} \tab  -\cr
 \tab \code{getNumerics} \tab  -\cr
 \tab \code{getReadablePath} \tab  -\cr
 \tab \code{getReadablePathname} \tab  -\cr
 \tab \code{getReadablePathnames} \tab  -\cr
 \tab \code{getRegularExpression} \tab  -\cr
 \tab \code{getVector} \tab  -\cr
 \tab \code{getVerbose} \tab  -\cr
 \tab \code{getWritablePath} \tab  -\cr
 \tab \code{getWritablePathname} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{classes}
\keyword{programming}
