%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  intToBin.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.binmode}
\alias{as.character.binmode}


\title{Converts a binary/octal/hexadecimal number into a string}

\description{
  Converts a binary/octal/hexadecimal number into a string.
}

\usage{
 \method{as.character}{binmode}(x, ...)
}

\arguments{
 \item{x}{Object to be converted.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{as.character.octmode()}, cf. \code{\link[base]{octmode}}.
  \code{\link{intToBin}}() (incl. \code{intToOct()} and \code{intToHex()}).
}




\keyword{manip}
\keyword{character}
\keyword{programming}
