%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.double.Verbose}
\alias{as.double.Verbose}
\alias{Verbose.as.double}
\alias{as.double,Verbose-method}

 \title{Gets a numeric value of this object}

 \description{
   Gets a numeric value of this object.  Returns what \code{\link[R.utils:getThreshold.Verbose]{*getThreshold}()} returns.
 }

 \usage{
\method{as.double}{Verbose}(x, ...)
}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{numeric}} value.
 }

 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[R.utils:getThreshold.Verbose]{*getThreshold}()} and \code{\link[R.utils:getThreshold.Verbose]{*getThreshold}()}.
   For more information see \code{\link{Verbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
