%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 12:00:04 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.logical.Verbose}
\alias{as.logical.Verbose}
\alias{Verbose.as.logical}
\alias{as.logical.Verbose}
\alias{as.logical,Verbose-method}


 \title{Gets a logical value of this object}

 \description{
   Gets a logical value of this object.  Returns \code{isVisible(this, level=this$defaultLevel)}.
 }

 \usage{\method{as.logical}{Verbose}(x, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{logical}} value.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   \code{\link[R.utils:isVisible.Verbose]{*isVisible}()}.
   For more information see \code{\link{Verbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
