%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Java.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asInt.Java}
\alias{asInt.Java}
\alias{Java.asInt}
\alias{asInt.Java}
\alias{asInt,Java-method}


\title{Converts an numeric to a Java integer}

\description{
  Converts an numeric to a Java integer.
}

\synopsis{asInt.Java(static, x, ...)}
\usage{Java$asInt(x, ...)}

\arguments{
 \item{x}{A \code{\link[base]{numeric}} \code{\link[base]{vector}}.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Java}}.
}
\keyword{internal}
\keyword{methods}
