%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SmartComments.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{convertComment.SmartComments}
\alias{convertComment.SmartComments}
\alias{SmartComments.convertComment}
\alias{convertComment,SmartComments-method}

 \title{Converts a single smart comment to R code}

 \description{
   Converts a single smart comment to R code.
 }

 \usage{
\method{convertComment}{SmartComments}(...)
}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Should return single \code{\link[base]{character}} of valid \R code.
 }

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{SmartComments}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
