%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Options.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{equals.Options}
\alias{equals.Options}
\alias{Options.equals}
\alias{equals.Options}
\alias{equals,Options-method}


 \title{Checks if this object is equal to another Options object}

 \description{
  Checks if this object is equal to another Options object.
 }

 \usage{\method{equals}{Options}(this, other, ...)}

 \arguments{
   \item{other}{Another Options object.}
   \item{...}{Not used.}
 }

 \value{Returns \code{\link[base:logical]{TRUE}} if they are equal, otherwise \code{\link[base:logical]{FALSE}}.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
