%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{equals.Verbose}
\alias{equals.Verbose}
\alias{Verbose.equals}
\alias{equals,Verbose-method}

 \title{Checks if this object is equal to another}

 \description{
  Checks if this object is equal to another.
 }

 \usage{
\method{equals}{Verbose}(this, other, ...)
}

 \arguments{
   \item{other}{Another Object.}
   \item{...}{Not used.}
 }

 \value{Returns \code{\link[base:logical]{TRUE}} if they are equal, otherwise \code{\link[base:logical]{FALSE}}.}

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
