%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{GString$getBuiltinDate}
\alias{GString$getBuiltinDate}
\alias{getBuiltinDate.GString}
\alias{GString.getBuiltinDate}
\alias{getBuiltinDate,GString-method}

\title{Gets the current date}

\description{
  Gets the current date.
}

\usage{
## Static method (use this):
## GString$getBuiltinDate(format="\%Y-\%m-\%d", ...)

## Don't use the below:
\method{getBuiltinDate}{GString}(static, format="\%Y-\%m-\%d", ...)
}

\arguments{
 \item{format}{A \code{\link[base]{character}} format string.}
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
