%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{GString$getBuiltinHostname}
\alias{GString$getBuiltinHostname}
\alias{getBuiltinHostname.GString}
\alias{GString.getBuiltinHostname}
\alias{getBuiltinHostname,GString-method}

\title{Gets the hostname of the system running R}

\description{
  Gets the hostname of the system running R.
}

\usage{
## Static method (use this):
## GString$getBuiltinHostname(...)

## Don't use the below:
\method{getBuiltinHostname}{GString}(static, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
