%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBuiltinHostname.GString}
\alias{getBuiltinHostname.GString}
\alias{GString.getBuiltinHostname}
\alias{getBuiltinHostname.GString}
\alias{getBuiltinHostname,GString-method}


\title{Gets the hostname of the system running R}

\description{
  Gets the hostname of the system running R.
}

\synopsis{getBuiltinHostname.GString(static, ...)}
\usage{GString$getBuiltinHostname(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
