%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBuiltinOs.GString}
\alias{getBuiltinOs.GString}
\alias{GString.getBuiltinOs}
\alias{getBuiltinOs.GString}
\alias{getBuiltinOs,GString-method}


\title{Gets the operating system of the running machine}

\description{
  Gets the operating system of the running machine.
}

\usage{
## Static method (use this):
## GString$getBuiltinOs(...)

## Don't use the below:
\method{getBuiltinOs}{GString}(static, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
