%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBuiltinPid.GString}
\alias{getBuiltinPid.GString}
\alias{GString.getBuiltinPid}
\alias{getBuiltinPid.GString}
\alias{getBuiltinPid,GString-method}


\title{Gets the process id of the current R session}

\description{
  Gets the process id of the current R session.
}

\synopsis{getBuiltinPid.GString(static, ...)}
\usage{GString$getBuiltinPid(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
