%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{GString$getBuiltinPid}
\alias{GString$getBuiltinPid}
\alias{getBuiltinPid.GString}
\alias{GString.getBuiltinPid}
\alias{getBuiltinPid,GString-method}

\title{Gets the process id of the current R session}

\description{
  Gets the process id of the current R session.
}

\usage{
## Static method (use this):
## GString$getBuiltinPid(...)

## Don't use the below:
\method{getBuiltinPid}{GString}(static, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
