%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBuiltinRhome.GString}
\alias{getBuiltinRhome.GString}
\alias{GString.getBuiltinRhome}
\alias{getBuiltinRhome.GString}
\alias{getBuiltinRhome,GString-method}


\title{Gets the path where R is installed}

\description{
  Gets the path where R is installed.
}

\synopsis{getBuiltinRhome.GString(static, ...)}
\usage{GString$getBuiltinRhome(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
