%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{GString$getBuiltinRhome}
\alias{GString$getBuiltinRhome}
\alias{getBuiltinRhome.GString}
\alias{GString.getBuiltinRhome}
\alias{getBuiltinRhome,GString-method}

\title{Gets the path where R is installed}

\description{
  Gets the path where R is installed.
}

\usage{
## Static method (use this):
## GString$getBuiltinRhome(...)

## Don't use the below:
\method{getBuiltinRhome}{GString}(static, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
