%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBuiltinRversion.GString}
\alias{getBuiltinRversion.GString}
\alias{GString.getBuiltinRversion}
\alias{getBuiltinRversion.GString}
\alias{getBuiltinRversion,GString-method}


\title{Gets the current R version}

\description{
  Gets the current R version.
}

\synopsis{getBuiltinRversion.GString(static, ...)}
\usage{GString$getBuiltinRversion(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
