%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{GString$getBuiltinRversion}
\alias{GString$getBuiltinRversion}
\alias{getBuiltinRversion.GString}
\alias{GString.getBuiltinRversion}
\alias{getBuiltinRversion,GString-method}

\title{Gets the current R version}

\description{
  Gets the current R version.
}

\usage{
## Static method (use this):
## GString$getBuiltinRversion(...)

## Don't use the below:
\method{getBuiltinRversion}{GString}(static, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
