%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{GString$getBuiltinUsername}
\alias{GString$getBuiltinUsername}
\alias{getBuiltinUsername.GString}
\alias{GString.getBuiltinUsername}
\alias{getBuiltinUsername,GString-method}

\title{Gets the username of the user running R}

\description{
  Gets the username of the user running R.
}

\usage{
## Static method (use this):
## GString$getBuiltinUsername(...)

## Don't use the below:
\method{getBuiltinUsername}{GString}(static, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
