%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Arguments.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getIndices.Arguments}
\alias{getIndices.Arguments}
\alias{Arguments.getIndices}
\alias{getIndices.Arguments}
\alias{getIndices,Arguments-method}

\alias{Arguments.getIndex}
\alias{getIndex.Arguments}
\alias{getIndex,Arguments-method}


\title{Coerces to a integer vector and validates}

\description{
 Coerces to a integer vector and validates.
}

\synopsis{getIndices.Arguments(static, ..., range=c(1, Inf))}
\usage{Arguments$getIndices(..., range=c(1, Inf))}

\arguments{
  \item{...}{Arguments passed to @method "getInteger".}
  \item{range}{Allowed range. See @method "getNumeric" for details.}
}

\value{
 Returns a \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Arguments}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
