%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Settings.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getLoadedPathname.Settings}
\alias{getLoadedPathname.Settings}
\alias{Settings.getLoadedPathname}
\alias{getLoadedPathname,Settings-method}

 \title{Gets the pathname of the settings file loaded}

 \description{
   Gets the pathname of the settings file loaded.
 }

 \usage{
\method{getLoadedPathname}{Settings}(this, ...)
}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns the absolute pathname (\code{\link[base]{character}} string) of the settings file
   loaded. If no file was read, \code{\link[base]{NULL}} is returned.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Settings}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
