%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTimestampFormat.Verbose}
\alias{getTimestampFormat.Verbose}
\alias{Verbose.getTimestampFormat}
\alias{getTimestampFormat,Verbose-method}

\title{Gets the default timestamp format}

\description{
  Gets the default timestamp format.
}

\usage{
\method{getTimestampFormat}{Verbose}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string or a \code{\link[base]{function}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.utils:setTimestampFormat.Verbose]{*setTimestampFormat}()}.
  \code{\link[R.utils:timestampOn.Verbose]{*timestampOn}()}.
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
