%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  hasUrlProtocol.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasUrlProtocol}
\alias{hasUrlProtocol.default}
\alias{hasUrlProtocol}

\title{Checks if one or several pathnames has a URL protocol}

\description{
 Checks if one or several pathnames has a URL protocol.
}

\usage{
\method{hasUrlProtocol}{default}(pathname, ...)
}

\arguments{
  \item{pathname}{A \code{\link[base]{character}} \code{\link[base]{vector}}.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}



\keyword{IO}
\keyword{programming}
