%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:26:25 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{intToBin}
\alias{intToBin}

\alias{intToOct}
\alias{intToHex}

\title{Converts an integer to a binary/octal/hexadecimal number}

\description{
  Converts an integer to a binary/octal/hexadecimal number.
}

\usage{
 intToBin(x)
 intToOct(x)
 intToHex(x)
}

\arguments{
 \item{x}{An \code{\link[base]{integer}} to be converted.}
}

\value{
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}




\keyword{manip}
\keyword{character}
\keyword{programming}
