%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  intToBin.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{intToBin}
\alias{intToBin}

\alias{intToOct}
\alias{intToHex}

\title{Converts an integer to a binary/octal/hexadecimal number}

\description{
  Converts an integer to a binary/octal/hexadecimal number.
}

\usage{
 intToBin(x)
 intToOct(x)
 intToHex(x)
}

\arguments{
 \item{x}{An \code{\link[base]{integer}} to be converted.}
}

\value{
 Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}




\keyword{manip}
\keyword{character}
\keyword{programming}
