%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  isAbsolutePath.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isAbsolutePath}
\alias{isAbsolutePath.default}
\alias{isAbsolutePath}

 \title{Checks if this pathname is absolute}

 \description{
   Checks if this pathname is absolute.
 }

 \usage{
\method{isAbsolutePath}{default}(pathname, ...)
}

 \arguments{
   \item{pathname}{A \code{\link[base]{character}} string of the pathname to be checked.}
   \item{...}{Not used.}
 }

 \value{
  Returns a \code{\link[base:logical]{TRUE}} if the pathname is absolute, otherwise \code{\link[base:logical]{FALSE}}.
 }



 \author{Henrik Bengtsson}



\keyword{IO}
\keyword{programming}
