%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Settings.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isModified.Settings}
\alias{isModified.Settings}
\alias{Settings.isModified}
\alias{isModified,Settings-method}

 \title{Checks if settings has been modified compared to whats on file}

 \description{
   Checks if settings has been modified compared to whats on file.
 }

 \usage{
\method{isModified}{Settings}(this, ...)
}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns \code{\link[base:logical]{TRUE}} if settings have been modified since lasted loaded, or if
   they never have been loaded. Otherwise \code{\link[base:logical]{FALSE}} is returned.
 }

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{Settings}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
