%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  isPackageInstalled.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isPackageInstalled}
\alias{isPackageInstalled.default}
\alias{isPackageInstalled}

\title{Checks if a package is installed or not}

\description{
  Checks if a package is installed or not.
}

\usage{
\method{isPackageInstalled}{default}(package, ...)
}

\arguments{
 \item{package}{The name of the package.}
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{logical}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link{isPackageLoaded}}().
}



\keyword{utilities}
\keyword{package}
