%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Assert.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isScalar.Assert}
\alias{isScalar.Assert}
\alias{Assert.isScalar}
\alias{isScalar.Assert}
\alias{isScalar,Assert-method}


 \title{Static method asserting thatan object is a single value}

 \description{
   Static method asserting thatan object is a single value.
 }

 \synopsis{isScalar.Assert(static, x, ...)}
\usage{Assert$isScalar(x, ...)}

 \arguments{
  \item{x}{Object to be checked.}
  \item{...}{Not used.}
 }

 \value{
   Returns (invisibly) \code{\link[base:logical]{TRUE}}, or throws an exception.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Assert}}.
 }
\keyword{internal}
\keyword{methods}
