%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  isUrl.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isUrl}
\alias{isUrl.default}
\alias{isUrl}

\title{Checks if one or several pathnames is URLs}

\description{
 Checks if one or several pathnames is URLs.
}

\usage{\method{isUrl}{default}(pathname, ...)}

\arguments{
  \item{pathname}{A \code{\link[base]{character}} \code{\link[base]{vector}}.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{IO}
\keyword{programming}
