%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:22 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isVisible.NullVerbose}
\alias{isVisible.NullVerbose}
\alias{NullVerbose.isVisible}
\alias{isVisible.NullVerbose}
\alias{isVisible,NullVerbose-method}


 \title{Checks if a certain verbose level will be shown or not}

 \description{
   Checks if a certain verbose level will be shown or not.
 }

 \usage{\method{isVisible}{NullVerbose}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns always \code{\link[base:logical]{FALSE}}.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{NullVerbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
