%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  NullVerbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{isVisible.NullVerbose}
\alias{isVisible.NullVerbose}
\alias{NullVerbose.isVisible}
\alias{isVisible,NullVerbose-method}

 \title{Checks if a certain verbose level will be shown or not}

 \description{
   Checks if a certain verbose level will be shown or not.
 }

 \usage{
\method{isVisible}{NullVerbose}(this, ...)
}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns always \code{\link[base:logical]{FALSE}}.
 }

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{NullVerbose}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
