%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Options.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{names.Options}
\alias{names.Options}
\alias{Options.names}
\alias{names,Options-method}

 \title{Gets the full pathname of all (non-list) options}

 \description{
  Gets the full pathname of all (non-list) options.
 }

 \usage{
\method{names}{Options}(x, ...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns a \code{\link[base]{vector}} of \code{\link[base]{character}} strings.}

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
