%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:28 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfOptions.Options}
\alias{nbrOfOptions.Options}
\alias{Options.nbrOfOptions}
\alias{nbrOfOptions.Options}
\alias{nbrOfOptions,Options-method}


 \title{Gets the number of options set}

 \description{
  Gets the number of options set.
 }

 \usage{\method{nbrOfOptions}{Options}(this, ...)}

 \arguments{
   \item{...}{Not used.}
 }

 \value{Returns an \code{\link[base]{integer}}.}

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{Options}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
