%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{newline.Verbose}
\alias{newline.Verbose}
\alias{Verbose.newline}
\alias{newline.Verbose}
\alias{newline,Verbose-method}


\title{Writes one or several empty lines}

\description{
  Writes one or several empty lines.
}

\usage{
\method{newline}{Verbose}(this, n=1, ..., level=this$defaultLevel)
}

\arguments{
 \item{n}{The number of empty lines to write.}
 \item{...}{Not used.}
 \item{level}{A \code{\link[base]{numeric}} value to be compared to the threshold.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
