%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GString-class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.GString}
\alias{print.GString}
\alias{GString.print}
\alias{print.GString}
\alias{print,GString-method}


\title{Prints the processed GString}

\description{
  Prints the processed GString.
}

\usage{\method{print}{GString}(x, ...)}

\arguments{
 \item{...}{Arguments passed to \code{\link[base]{print}}().}
}

\value{
  Returns (invisibly) the process GString \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GString}}.
}
\keyword{internal}
\keyword{methods}
