%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProgressBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{reset.ProgressBar}
\alias{reset.ProgressBar}
\alias{ProgressBar.reset}
\alias{reset.ProgressBar}
\alias{reset,ProgressBar-method}


 \title{Reset progress bar}

 \description{
   Reset progress bar by setting the value to zero and updating the display.
 }

 \usage{\method{reset}{ProgressBar}(this, visual=TRUE, ...)}

 \arguments{
  \item{visual}{If \code{\link[base:logical]{TRUE}}, the progress bar is redraw, otherwise not.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[R.utils:setValue.ProgressBar]{*setValue}()}.
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
