%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 11:59:38 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{resetWarnings}
\alias{resetWarnings.default}
\alias{resetWarnings}

 \title{Resets recorded warnings}

 \description{
  Resets recorded warnings.
 }

 \usage{\method{resetWarnings}{default}(...)}

 \arguments{
   \item{...}{Not used.}
 }

 \value{
  Returns (invisibly) the number of warnings removed.
 }



 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   \code{\link[base]{warnings}}()
 }


\keyword{programming}
