%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  resetWarnings.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{resetWarnings}
\alias{resetWarnings.default}
\alias{resetWarnings}

 \title{Resets recorded warnings}

 \description{
  Resets recorded warnings.
 }

 \usage{
\method{resetWarnings}{default}(...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \value{
  Returns (invisibly) the number of warnings removed.
 }



 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[base]{warnings}}()
 }


\keyword{programming}
