%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:27:12 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setDefaultLevel.Verbose}
\alias{setDefaultLevel.Verbose}
\alias{Verbose.setDefaultLevel}
\alias{setDefaultLevel.Verbose}
\alias{setDefaultLevel,Verbose-method}


\title{Sets the current default verbose level}

\description{
  Sets the current default verbose level.
}

\usage{\method{setDefaultLevel}{Verbose}(this, level, ...)}

\arguments{
 \item{level}{A \code{\link[base]{numeric}} value.}
 \item{...}{Not used.}
}

\value{
  Returns old default level.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
