%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProgressBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setMaxValue.ProgressBar}
\alias{setMaxValue.ProgressBar}
\alias{ProgressBar.setMaxValue}
\alias{setMaxValue.ProgressBar}
\alias{setMaxValue,ProgressBar-method}


 \title{Sets maximum value}

 \description{
   Sets maximum value.
 }

 \usage{
\method{setMaxValue}{ProgressBar}(this, maxValue, ...)
}

 \arguments{
  \item{maxValue}{New maximum value.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
