%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProgressBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setStepLength.ProgressBar}
\alias{setStepLength.ProgressBar}
\alias{ProgressBar.setStepLength}
\alias{setStepLength.ProgressBar}
\alias{setStepLength,ProgressBar-method}


 \title{Sets default step length}

 \description{
   Sets default step length.
 }

 \usage{
\method{setStepLength}{ProgressBar}(this, stepLength, ...)
}

 \arguments{
  \item{stepLength}{New default step length.}
  \item{...}{Not used.}
 }

 \value{
   Returns on step length.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{ProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
