%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Verbose.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setThreshold.Verbose}
\alias{setThreshold.Verbose}
\alias{Verbose.setThreshold}
\alias{setThreshold.Verbose}
\alias{setThreshold,Verbose-method}


\title{Sets verbose threshold}

\description{
  Sets verbose threshold. Output requests below this threshold will be ignored.
}

\usage{\method{setThreshold}{Verbose}(this, threshold, ...)}

\arguments{
 \item{threshold}{A \code{\link[base]{numeric}} threshold.}
 \item{...}{Not used.}
}

\value{
  Returns old threshold.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.utils:getThreshold.Verbose]{*getThreshold}()} and \code{\link[R.utils:isVisible.Verbose]{*isVisible}()}.
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
