%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  System.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{System}
\docType{class}
\alias{System}


\title{Static class to query information about the system}

\description{
 Package:  R.utils \cr
\bold{Class System}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{System}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{System}\cr
extends \link[R.oo]{Object}\cr



 The System class contains several useful class fields and methods. It
 cannot be instantiated.
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{currentTimeMillis} \tab  -\cr
 \tab \code{findGhostscript} \tab  -\cr
 \tab \code{findGraphicsDevice} \tab  -\cr
 \tab \code{getHostname} \tab  -\cr
 \tab \code{getUsername} \tab  -\cr
 \tab \code{openBrowser} \tab  -\cr
 \tab \code{parseDebian} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
