%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Thu Jul 21 12:00:10 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{timestamp.Verbose}
\alias{timestamp.Verbose}
\alias{Verbose.timestamp}
\alias{timestamp.Verbose}
\alias{timestamp,Verbose-method}


\title{Writes a timestamp}

\description{
  Writes a timestamp with default format [2005-06-23 21:20:03].
}

\usage{\method{timestamp}{Verbose}(this, stamp=format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"), ...)}

\arguments{
 \item{stamp}{A timestamp \code{\link[base]{character}} to be written.}
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Verbose}}.
}


\keyword{internal}
\keyword{methods}
\keyword{programming}
