%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileProgressBar.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{update.FileProgressBar}
\alias{update.FileProgressBar}
\alias{FileProgressBar.update}
\alias{update,FileProgressBar-method}

 \title{Updates file progress bar}

 \usage{
\method{update}{FileProgressBar}(object, visual=TRUE, ...)
}

 \description{
   Updates file progress bar.
 }

 \arguments{
  \item{visual}{If \code{\link[base:logical]{TRUE}}, the file is resized according to the value of
    the progress bar, otherwise not.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{FileProgressBar}}.
 }
\keyword{internal}
\keyword{methods}
