%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SmartComments.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{validate.SmartComments}
\alias{validate.SmartComments}
\alias{SmartComments.validate}
\alias{validate,SmartComments-method}

 \title{Validates the compiled lines}

 \description{
   Validates the compiled lines
 }

 \usage{
\method{validate}{SmartComments}(this, lines, ...)
}

 \arguments{
  \item{lines}{A \code{\link[base]{character}} \code{\link[base]{vector}} of lines of code to validated.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson}

 \seealso{
   For more information see \code{\link{SmartComments}}.
 }


\keyword{internal}
\keyword{methods}
\keyword{programming}
